+++ CHANGELOG.md
@@ -4111,3 +4111,4 @@
 - [2026-09-01 12:46] test(forms): implement widget integration tests for forms
 - [2026-09-01 13:49] test(theme): implement widget integration tests for theme
 - [2026-09-01 14:55] feat(test): implement responsive test view component
+- [2026-09-01 15:06] chore(analytics): update dependencies and cleanup unused imports in analytics